+++ lib/vissen/input/message/pitch_bend_change.rb
@@ -25,7 +25,7 @@ module Vissen
         class << self
           # TODO: Check the range on value.
           def create(value = 0.0, **args)
-            bin_value = (value.to_f * CENTER_VALUE).to_i + CENTER_VALUE
+            bin_value = (value.to_f * CENTER_VALUE).round + CENTER_VALUE
 
             super(bin_value & 0xFF, bin_value >> 7, **args)
           end