+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Broker#run_once now returns true if a message was handled.
 - PitchBendChange.create now accepts a float value instead of the raw bytes.
 
+### Fixed
+- Bug in Message::Base.create that ment the status was always set to zero.
+
 ## [0.2.0] - 2018-04-11
 ### Added
 - Subscription object to handle objects subscribing to incoming messages.